requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
django-rest-framework==0.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
pytz==2018.9
